Add files via upload
Added the background, the ship and player movement. As of right now you could use the app to do a little etch a sketch (without the ability to stop) as the ship always leaves a trail of other ship in the background. Hopefully I can use this code to create an etch a sketch or tron like game later in the line.